fix for pool limit error

+++ basemodel.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Column,String,REAL,Text,Integer
+from sqlalchemy import Column,String,REAL,Text,BigInteger
 
 class templatemodel(object):
     Symbol = Column(String,index=True)
-    Timestamp = Column(Integer, index=True, unique = True,primary_key=True)
+    Timestamp = Column(BigInteger, index=True, unique = True,primary_key=True)
     Volume = Column(REAL)
     Open = Column(REAL)
     Close = Column(REAL)
